Generate stratified, block randomized allocation sequence

Stratified block randomization sequence with optional redcap API. Each stratum will have its own unique allocation sequence.

Usage

randomizR(
  strata,
  block.size,
  block.probabilities = NULL,
  replications = 50,
  allocation.levels = c("1", "2"),
  allocation.name = "redcap_randomization_group",
  block.delay = 10,
  token,
  seed = "none",
  report = T,
  print = F
)

Arguments

strata

named list of strata levels (e.g. list("age" = c(1,2), "sex" = c("f","m"))). If missing, an API token can be provided (see "token")

block.size

vector of block size(s).

block.probabilities

vector of block size probabilities

replications

length of the allocation sequence in each stratum

allocation.levels

vector of the allocated treatment levels, default = c("1", "2")

allocation.name

column name for the allocated treatment levels, default = "allocation"

block.delay

number of blocks that is fixed to the first block.size to avoid large blocks in the beginning of the sequence

token

token for redcap API to automatically extract strata levels from redcap. If strata is a vector of variables, token must be provided.

seed

seed for reproducibility. If "none", no seed is set

report

whether the allocation report should be made

print

whether a full report for each stratum should be printed

Value

a data frame containing rows equal to the product of unique strata and replications.

Examples

allocation <-
  randomizR(strata=c("age_group_depth",
                     "gender_depth",
                     "ana_site_depth",
                     "immuno_depth",
                     "seq_width_depth",
                     "seq_slnb_depth"),
            block.size = c(2,4),
            block.probabilities = c(0.9,0.1),
            replications = 50,
            block.delay = 10,
            token = '859A8159EB61A5AD6ABB8DE0BC78E67F',
            seed = 2,
            report = T,
            print=T)
#> Stratified block randomization complete
#> 
#> Stratified on 6 strata with the levels:
#> $age_group_depth
#> [1] "1" "2"
#> 
#> $gender_depth
#> [1] "1" "2"
#> 
#> $ana_site_depth
#> [1] "1" "2"
#> 
#> $immuno_depth
#> [1] "1" "2"
#> 
#> $seq_width_depth
#> [1] "1" "2" "3"
#> 
#> $seq_slnb_depth
#> [1] "1" "2" "3"
#> 
#> Number of unique strata: 144
#> 
#> Strata balances:
#> # A tibble: 5 × 2
#>   `2(%)` `4(%)`
#>    <dbl>  <dbl>
#> 1   75      4.2
#> 2   86.4    4.2
#> 3   91.3    8.7
#> 4   95.8   13.6
#> 5   95.8   25  
#> # A tibble: 144 × 9
#>     Stratum    Replications Allocation.ratio Block.Count Block.Sizes   `2`   `4`
#>     <chr>             <int> <chr>                  <int> <chr>       <int> <int>
#>   1 1 1 1 1 1…           50 25/25                     24 2,4            23     1
#>   2 2 1 1 1 1…           50 25/25                     21 2,4            17     4
#>   3 1 2 1 1 1…           50 25/25                     22 2,4            19     3
#>   4 2 2 1 1 1…           50 25/25                     21 2,4            17     4
#>   5 1 1 2 1 1…           50 25/25                     23 2,4            21     2
#>   6 2 1 2 1 1…           50 25/25                     23 2,4            21     2
#>   7 1 2 2 1 1…           50 25/25                     22 2,4            19     3
#>   8 2 2 2 1 1…           50 25/25                     24 2,4            23     1
#>   9 1 1 1 2 1…           50 25/25                     23 2,4            21     2
#>  10 2 1 1 2 1…           50 25/25                     23 2,4            21     2
#>  11 1 2 1 2 1…           50 25/25                     24 2,4            23     1
#>  12 2 2 1 2 1…           50 25/25                     22 2,4            19     3
#>  13 1 1 2 2 1…           50 25/25                     23 2,4            21     2
#>  14 2 1 2 2 1…           50 25/25                     24 2,4            23     1
#>  15 1 2 2 2 1…           50 25/25                     23 2,4            21     2
#>  16 2 2 2 2 1…           50 25/25                     23 2,4            21     2
#>  17 1 1 1 1 2…           50 25/25                     22 2,4            19     3
#>  18 2 1 1 1 2…           50 25/25                     23 2,4            21     2
#>  19 1 2 1 1 2…           50 25/25                     21 2,4            17     4
#>  20 2 2 1 1 2…           50 25/25                     23 2,4            21     2
#>  21 1 1 2 1 2…           50 25/25                     22 2,4            19     3
#>  22 2 1 2 1 2…           50 25/25                     23 2,4            21     2
#>  23 1 2 2 1 2…           50 25/25                     22 2,4            19     3
#>  24 2 2 2 1 2…           50 25/25                     24 2,4            23     1
#>  25 1 1 1 2 2…           50 25/25                     23 2,4            21     2
#>  26 2 1 1 2 2…           50 25/25                     23 2,4            21     2
#>  27 1 2 1 2 2…           50 25/25                     23 2,4            21     2
#>  28 2 2 1 2 2…           50 25/25                     22 2,4            19     3
#>  29 1 1 2 2 2…           50 25/25                     23 2,4            21     2
#>  30 2 1 2 2 2…           50 25/25                     23 2,4            21     2
#>  31 1 2 2 2 2…           50 25/25                     23 2,4            21     2
#>  32 2 2 2 2 2…           50 25/25                     24 2,4            23     1
#>  33 1 1 1 1 3…           50 25/25                     21 2,4            17     4
#>  34 2 1 1 1 3…           50 25/25                     24 2,4            23     1
#>  35 1 2 1 1 3…           50 25/25                     23 2,4            21     2
#>  36 2 2 1 1 3…           50 25/25                     24 2,4            23     1
#>  37 1 1 2 1 3…           50 25/25                     24 2,4            23     1
#>  38 2 1 2 1 3…           50 25/25                     22 2,4            19     3
#>  39 1 2 2 1 3…           50 25/25                     22 2,4            19     3
#>  40 2 2 2 1 3…           50 25/25                     22 2,4            19     3
#>  41 1 1 1 2 3…           50 25/25                     24 2,4            23     1
#>  42 2 1 1 2 3…           50 25/25                     24 2,4            23     1
#>  43 1 2 1 2 3…           50 25/25                     24 2,4            23     1
#>  44 2 2 1 2 3…           50 25/25                     22 2,4            19     3
#>  45 1 1 2 2 3…           50 25/25                     23 2,4            21     2
#>  46 2 1 2 2 3…           50 25/25                     22 2,4            19     3
#>  47 1 2 2 2 3…           50 25/25                     23 2,4            21     2
#>  48 2 2 2 2 3…           50 25/25                     24 2,4            23     1
#>  49 1 1 1 1 1…           50 25/25                     23 2,4            21     2
#>  50 2 1 1 1 1…           50 25/25                     23 2,4            21     2
#>  51 1 2 1 1 1…           50 25/25                     23 2,4            21     2
#>  52 2 2 1 1 1…           50 25/25                     22 2,4            19     3
#>  53 1 1 2 1 1…           50 25/25                     23 2,4            21     2
#>  54 2 1 2 1 1…           50 25/25                     24 2,4            23     1
#>  55 1 2 2 1 1…           50 25/25                     22 2,4            19     3
#>  56 2 2 2 1 1…           50 25/25                     23 2,4            21     2
#>  57 1 1 1 2 1…           50 25/25                     22 2,4            19     3
#>  58 2 1 1 2 1…           50 25/25                     24 2,4            23     1
#>  59 1 2 1 2 1…           50 25/25                     23 2,4            21     2
#>  60 2 2 1 2 1…           50 25/25                     23 2,4            21     2
#>  61 1 1 2 2 1…           50 25/25                     24 2,4            23     1
#>  62 2 1 2 2 1…           50 25/25                     22 2,4            19     3
#>  63 1 2 2 2 1…           50 25/25                     23 2,4            21     2
#>  64 2 2 2 2 1…           50 25/25                     23 2,4            21     2
#>  65 1 1 1 1 2…           50 25/25                     23 2,4            21     2
#>  66 2 1 1 1 2…           50 25/25                     24 2,4            23     1
#>  67 1 2 1 1 2…           50 25/25                     24 2,4            23     1
#>  68 2 2 1 1 2…           50 25/25                     22 2,4            19     3
#>  69 1 1 2 1 2…           50 25/25                     24 2,4            23     1
#>  70 2 1 2 1 2…           50 25/25                     23 2,4            21     2
#>  71 1 2 2 1 2…           50 25/25                     21 2,4            17     4
#>  72 2 2 2 1 2…           50 25/25                     22 2,4            19     3
#>  73 1 1 1 2 2…           50 25/25                     24 2,4            23     1
#>  74 2 1 1 2 2…           50 25/25                     24 2,4            23     1
#>  75 1 2 1 2 2…           50 25/25                     23 2,4            21     2
#>  76 2 2 1 2 2…           50 25/25                     24 2,4            23     1
#>  77 1 1 2 2 2…           50 25/25                     23 2,4            21     2
#>  78 2 1 2 2 2…           50 25/25                     24 2,4            23     1
#>  79 1 2 2 2 2…           50 25/25                     24 2,4            23     1
#>  80 2 2 2 2 2…           50 25/25                     23 2,4            21     2
#>  81 1 1 1 1 3…           50 25/25                     21 2,4            17     4
#>  82 2 1 1 1 3…           50 25/25                     23 2,4            21     2
#>  83 1 2 1 1 3…           50 25/25                     22 2,4            19     3
#>  84 2 2 1 1 3…           50 25/25                     23 2,4            21     2
#>  85 1 1 2 1 3…           50 25/25                     24 2,4            23     1
#>  86 2 1 2 1 3…           50 25/25                     22 2,4            19     3
#>  87 1 2 2 1 3…           50 25/25                     21 2,4            17     4
#>  88 2 2 2 1 3…           50 25/25                     24 2,4            23     1
#>  89 1 1 1 2 3…           50 25/25                     21 2,4            17     4
#>  90 2 1 1 2 3…           50 25/25                     23 2,4            21     2
#>  91 1 2 1 2 3…           50 25/25                     22 2,4            19     3
#>  92 2 2 1 2 3…           50 25/25                     23 2,4            21     2
#>  93 1 1 2 2 3…           50 25/25                     24 2,4            23     1
#>  94 2 1 2 2 3…           50 25/25                     20 2,4            15     5
#>  95 1 2 2 2 3…           50 25/25                     22 2,4            19     3
#>  96 2 2 2 2 3…           50 25/25                     24 2,4            23     1
#>  97 1 1 1 1 1…           50 25/25                     23 2,4            21     2
#>  98 2 1 1 1 1…           50 25/25                     24 2,4            23     1
#>  99 1 2 1 1 1…           50 25/25                     23 2,4            21     2
#> 100 2 2 1 1 1…           50 25/25                     23 2,4            21     2
#> 101 1 1 2 1 1…           50 25/25                     24 2,4            23     1
#> 102 2 1 2 1 1…           50 25/25                     23 2,4            21     2
#> 103 1 2 2 1 1…           50 25/25                     22 2,4            19     3
#> 104 2 2 2 1 1…           50 25/25                     23 2,4            21     2
#> 105 1 1 1 2 1…           50 25/25                     23 2,4            21     2
#> 106 2 1 1 2 1…           50 25/25                     24 2,4            23     1
#> 107 1 2 1 2 1…           50 25/25                     21 2,4            17     4
#> 108 2 2 1 2 1…           50 25/25                     21 2,4            17     4
#> 109 1 1 2 2 1…           50 25/25                     24 2,4            23     1
#> 110 2 1 2 2 1…           50 25/25                     21 2,4            17     4
#> 111 1 2 2 2 1…           50 25/25                     23 2,4            21     2
#> 112 2 2 2 2 1…           50 25/25                     23 2,4            21     2
#> 113 1 1 1 1 2…           50 25/25                     23 2,4            21     2
#> 114 2 1 1 1 2…           50 25/25                     22 2,4            19     3
#> 115 1 2 1 1 2…           50 25/25                     24 2,4            23     1
#> 116 2 2 1 1 2…           50 25/25                     22 2,4            19     3
#> 117 1 1 2 1 2…           50 25/25                     23 2,4            21     2
#> 118 2 1 2 1 2…           50 25/25                     21 2,4            17     4
#> 119 1 2 2 1 2…           50 25/25                     24 2,4            23     1
#> 120 2 2 2 1 2…           50 25/25                     24 2,4            23     1
#> 121 1 1 1 2 2…           50 25/25                     23 2,4            21     2
#> 122 2 1 1 2 2…           50 25/25                     23 2,4            21     2
#> 123 1 2 1 2 2…           50 25/25                     23 2,4            21     2
#> 124 2 2 1 2 2…           50 25/25                     22 2,4            19     3
#> 125 1 1 2 2 2…           50 25/25                     24 2,4            23     1
#> 126 2 1 2 2 2…           50 25/25                     22 2,4            19     3
#> 127 1 2 2 2 2…           50 25/25                     24 2,4            23     1
#> 128 2 2 2 2 2…           50 25/25                     23 2,4            21     2
#> 129 1 1 1 1 3…           50 25/25                     24 2,4            23     1
#> 130 2 1 1 1 3…           50 25/25                     23 2,4            21     2
#> 131 1 2 1 1 3…           50 25/25                     24 2,4            23     1
#> 132 2 2 1 1 3…           50 25/25                     23 2,4            21     2
#> 133 1 1 2 1 3…           50 25/25                     23 2,4            21     2
#> 134 2 1 2 1 3…           50 25/25                     23 2,4            21     2
#> 135 1 2 2 1 3…           50 25/25                     22 2,4            19     3
#> 136 2 2 2 1 3…           50 25/25                     24 2,4            23     1
#> 137 1 1 1 2 3…           50 25/25                     24 2,4            23     1
#> 138 2 1 1 2 3…           50 25/25                     21 2,4            17     4
#> 139 1 2 1 2 3…           50 25/25                     24 2,4            23     1
#> 140 2 2 1 2 3…           50 25/25                     23 2,4            21     2
#> 141 1 1 2 2 3…           50 25/25                     21 2,4            17     4
#> 142 2 1 2 2 3…           50 25/25                     24 2,4            23     1
#> 143 1 2 2 2 3…           50 25/25                     20 2,4            15     5
#> 144 2 2 2 2 3…           50 25/25                     23 2,4            21     2
#> # ℹ 2 more variables: `2(%)` <dbl>, `4(%)` <dbl>